feat(practica04): :construction: Antonio Gil · añadir avances 07/May

+++ AntonioGil/practica04/src/main/java/com/viewnex/bsan/practica04/entity/Actor.java
@@ -25,7 +25,7 @@ public class Actor {
     String name;
 
     @Column(name = "ACTOR_AGE")
-    short age;
+    int age;
 
     @Column(name = "ACTOR_NATIONALITY")
     String nationality;
