+++ AntonioGil/practica04/src/main/java/com/viewnex/bsan/practica04/entity/ProductionCompany.java
@@ -25,7 +25,7 @@ public class ProductionCompany {
     String name;
 
     @Column(name = "PRODUCTION_COMPANY_YEAR_FOUNDED")
-    short yearFounded;
+    int yearFounded;
 
     @Override
     public boolean equals(Object o) {
